pyproject.toml
[project]
name = "vocoders"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "reppy4620", email = "ilikeniku@icloud.com" }
]
dependencies = [
    "lightning>=2.3.3",
    "torchaudio>=2.3.1",
    "joblib>=1.4.2",
    "pyworld>=0.3.4",
    "soundfile>=0.12.1",
    "pesq>=0.0.4",
    "hydra-core>=1.3.2",
    "matplotlib>=3.9.1",
    "pandas>=2.2.2",
    "scipy>=1.14.0",
    "loguru>=0.7.2",
    "nnaudio>=0.3.3",
    "tensorboard>=2.17.0",
    "librosa>=0.10.2.post1",
]
readme = "README.md"
requires-python = ">= 3.11"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/vocoders"]
